+++ plugins/synced_folders/nfs/synced_folder.rb
@@ -37,6 +37,16 @@ module VagrantPlugins
         raise Vagrant::Errors::NFSNoHostIP if !nfsopts[:nfs_host_ip]
         raise Vagrant::Errors::NFSNoGuestIP if !nfsopts[:nfs_machine_ip]
 
+        if machine.guest.capability?(:nfs_client_installed)
+          installed = machine.guest.capability(:nfs_client_installed)
+          if !installed
+            can_install = machine.guest.capability?(:nfs_client_install)
+            raise Vagrant::Errors::NFSClientNotInstalledInGuest if !can_install
+            machine.ui.info I18n.t("vagrant.actions.vm.nfs.installing")
+            machine.guest.capability(:nfs_client_install)
+          end
+        end
+
         machine_ip = nfsopts[:nfs_machine_ip]
         machine_ip = [machine_ip] if !machine_ip.is_a?(Array)